migrate to mysql

+++ src/main/resources/db/migration/V1.7__0_0_rules_deleted.sql
@@ -1,1 +1,1 @@
-ALTER TABLE rules ADD `deleted` INTEGER DEFAULT 0;
+ALTER TABLE rules ADD `deleted` TINYINT DEFAULT 0;
