+++ src/main/resources/db/migration/V1.8__0_0_rules_priority.sql
@@ -1,3 +1,3 @@
-ALTER TABLE rules ADD `priority` INTEGER DEFAULT 0;
+ALTER TABLE rules ADD `priority` SMALLINT DEFAULT 0;
 
-ALTER TABLE rules ADD `stop_process_rules` INTEGER DEFAULT 0;
+ALTER TABLE rules ADD `stop_process_rules` TINYINT DEFAULT 0;
